pyproject.toml
[tool.poetry]
name = "bot-sobes-tg"
version = "0.1.0"
description = ""
authors = ["ya_novikov <arhdomain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
sqladmin = {extras = ["all"], version = "^0.20.1"}
passlib = "^1.7.4"
pyjwt = "^2.10.1"
sqlalchemy-utils = "^0.41.2"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.test.dependencies]
fakeredis = "2.26.2"
pytest = "8.3.4"
pytest-asyncio = "0.25.0"
pytest-env = "1.1.5"
pytest-mock = "3.14.0"
pytest-order = "1.3.0"

[tool.poetry.group.dev.dependencies]
bandit = "1.8.0"
black = "24.10.0"
ruff = "0.8.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = ["secrets/.env"]
env_override_existing_values = 1
python_classes = "*Test"
python_files = "tests.py test_*.py"
pythonpath = "src"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope="session"

[tool.black]
extend-exclude = 'migrations'
include = '\.pyi?$'
line-length = 88
target-version = ['py38', 'py311']

[tool.bandit]
exclude_dirs = ["src/migrations", "venv"]
skips = ["B101"]
targets = ["src", "tests"]

[tool.ruff]
lint.ignore = ["E712"]
lint.select = [
  "E", # pycodestyle
  "F", # Pyflakes
  "UP", # pyupgrade
  "B", # flake8-bugbear
  "SIM", # flake8-simplify
  "I", # isort
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = [
  "A",
  "B",
  "C",
  "D",
  "E",
  "F",
  "G",
  "I",
  "N",
  "Q",
  "S",
  "T",
  "W",
  "ANN",
  "ARG",
  "BLE",
  "COM",
  "DJ",
  "DTZ",
  "EM",
  "ERA",
  "EXE",
  "FBT",
  "ICN",
  "INP",
  "ISC",
  "NPY",
  "PD",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "TCH",
  "TID",
  "TRY",
  "UP",
  "YTT",
]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "*migrations*",
  "app/app/management/commands/*",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 18

[tool.ruff.lint.isort]
lines-after-imports = 2
lines-between-types = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
